fix: move pages, lib directory
