+++ src/app/pages/about/about.component.scss
