+++ src/app/pages/about/about.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { AuthService } from 'src/app/services/auth.service';
 import { AuthServiceStub } from 'src/test/service.stub';
-import { AboutComponent } from './about.component';
+import AboutComponent from './about.component';
 
 describe('AboutComponent', () => {
   let component: AboutComponent;
