+++ src/app/pages/article-detail/article-detail.component.scss
