+++ src/app/pages/article-detail/article-detail.component.ts
@@ -28,8 +28,8 @@ import { ViewCountService } from 'src/app/services/view-count.service';
 import { SafeHTMLPipe } from '../../pipes/safe-html.pipe';
 import { EncodeUrlPipe } from '../../pipes/encode-url.pipe';
 import { StringToLinkPipe } from '../../pipes/string-to-link.pipe';
-import { RecommendArticleComponent } from '../recommend-article/recommend-article.component';
-import { ArticleCommentComponent } from '../article-comment/article-comment.component';
+import { RecommendArticleComponent } from '../../mypage/recommend-article/recommend-article.component';
+import { ArticleCommentComponent } from '../../mypage/article-comment/article-comment.component';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatLegacyChipsModule } from '@angular/material/legacy-chips';
 import { ArticleEditButtonsComponent } from '../../shared-article-edit/article-edit-buttons/article-edit-buttons.component';
