+++ src/app/pages/check/check.component.ts
