+++ src/app/pages/create/create.component.spec.ts
@@ -15,7 +15,7 @@ import { ArticleService } from 'src/app/services/article.service';
 import { AuthService } from 'src/app/services/auth.service';
 import { FireAnalyticsStub, FirestoreStub } from 'src/test/firebase.stub';
 import { ArticleServiceStub, AuthServiceStub } from 'src/test/service.stub';
-import { CreateComponent } from './create.component';
+import CreateComponent from './create.component';
 
 describe('CreateComponent', () => {
   let component: CreateComponent;
