+++ src/app/pages/create/create.component.ts
@@ -17,8 +17,8 @@ import { SeoService } from 'src/app/services/seo.service';
 import { UserData } from '@interfaces/user';
 import { AngularFirestore } from '@angular/fire/compat/firestore';
 import { AngularFireAnalytics } from '@angular/fire/compat/analytics';
-import { EditorComponent } from '../editor/editor.component';
-import { TagFormComponent } from '../tag-form/tag-form.component';
+import { EditorComponent } from '../../articles/editor/editor.component';
+import { TagFormComponent } from '../../articles/tag-form/tag-form.component';
 import { MatLegacyInputModule } from '@angular/material/legacy-input';
 import { MatLegacyFormFieldModule } from '@angular/material/legacy-form-field';
 import { MatLegacySlideToggleModule } from '@angular/material/legacy-slide-toggle';
