+++ src/app/pages/liked-articles/liked-articles.component.scss
