+++ src/app/pages/mypage/mypage.component.ts
