+++ src/app/pages/privacy/privacy.component.ts
