+++ src/app/pages/search-result/search-result.component.spec.ts
