+++ src/app/pages/settings/settings.component.spec.ts
@@ -8,7 +8,7 @@ import {
 import { AuthService } from 'src/app/services/auth.service';
 import { UserService } from 'src/app/services/user.service';
 import { AuthServiceStub, UserServiceStub } from 'src/test/service.stub';
-import { SettingsComponent } from './settings.component';
+import SettingsComponent from './settings.component';
 
 describe('SettingsComponent', () => {
   let component: SettingsComponent;
