+++ src/app/pages/terms/terms.component.html
