+++ src/app/pages/terms/terms.component.scss
