+++ src/app/pages/terms/terms.component.spec.ts
