+++ src/app/pages/terms/terms.component.ts
