+++ src/app/pages/top/top.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ArticleService } from 'src/app/services/article.service';
 import { AuthService } from 'src/app/services/auth.service';
 import { ArticleServiceStub, AuthServiceStub } from 'src/test/service.stub';
-import { TopComponent } from './top.component';
+import TopComponent from './top.component';
 
 describe('TopComponent', () => {
   let component: TopComponent;
