+++ src/app/pages/top/top.component.ts
@@ -4,11 +4,11 @@ import { Observable, of } from 'rxjs';
 import { ArticleWithAuthor } from 'functions/src/interfaces/article-with-author';
 import { take } from 'rxjs/operators';
 import { SeoService } from 'src/app/services/seo.service';
-import { ArticleCardSkeltonComponent } from '../shared-article-card/article-card-skelton/article-card-skelton.component';
-import { ArticleCardComponent } from '../shared-article-card/article-card/article-card.component';
+import { ArticleCardSkeltonComponent } from '../../shared-article-card/article-card-skelton/article-card-skelton.component';
+import { ArticleCardComponent } from '../../shared-article-card/article-card/article-card.component';
 import { NgIf, NgFor, AsyncPipe } from '@angular/common';
 import { MatIconModule } from '@angular/material/icon';
-import { WelcomeComponent } from '../welcome/welcome.component';
+import { WelcomeComponent } from '../../welcome/welcome.component';
 
 @Component({
   selector: 'app-top',
